+++ application/modules/editor/Plugins/MatchAnalysis/Analysis.php
@@ -262,7 +262,18 @@ class editor_Plugins_MatchAnalysis_Analysis extends editor_Plugins_MatchAnalysis
 
             return $bestResult;
         }
-        $masterHasResult = ! empty($this->repetitionByHash[$segmentHash]);
+        $masterResult = $this->repetitionByHash[$segmentHash];
+        $masterHasResult = ! empty($masterResult);
+
+        // If it is repetition match, no need to do anything bellow. Just pre-translate the segment out of the master
+        // repetition
+        if ($masterResult->matchrate === FileBasedInterface::TERMCOLLECTION_MATCH_VALUE) {
+            $this->saveAnalysis($segment, FileBasedInterface::REPETITION_MATCH_VALUE, 0);
+            $masterResult->matchrate = FileBasedInterface::REPETITION_MATCH_VALUE;
+            $masterResult->isRepetition = true;
+
+            return $masterResult;
+        }
 
         // DESCRIPTION BEHAVIOUR FOR REPETITIONS
         // for the analysis, a repetition is always counted as 102% match!